+++ sdk/go/x/auto/workspace.go
@@ -56,6 +56,17 @@ type Workspace interface {
 	RemoveAllConfig(context.Context, string, []string) error
 	// RefreshConfig gets and sets the config map used with the last Update for Stack matching fullyQualifiedStackName.
 	RefreshConfig(context.Context, string) (ConfigMap, error)
+	// GetEnvVars returns the environment values scoped to the current workspace.
+	GetEnvVars() map[string]string
+	// SetEnvVars sets the specified map of environment values scoped to the current workspace.
+	// These values will be passed to all Workspace and Stack level commands.
+	SetEnvVars(map[string]string) error
+	// SetEnvVar sets the specified environment value scoped to the current workspace.
+	// This value will be passed to all Workspace and Stack level commands.
+	SetEnvVar(string, string)
+	// UnsetEnvVar unsets the specified environment value scoped to the current workspace.
+	// This value will be removed from all Workspace and Stack level commands.
+	UnsetEnvVar(string)
 	// WorkDir returns the working directory to run Pulumi CLI commands.
 	WorkDir() string
 	// PulumiHome returns the directory override for CLI metadata if set.